Extract class MockRedirectHandler
