+++ client/trino-client/src/test/java/io/trino/client/auth/external/MockRedirectHandler.java
@@ -1,0 +1,34 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.client.auth.external;
+
+import java.net.URI;
+
+public class MockRedirectHandler
+        implements RedirectHandler
+{
+    private URI redirectedTo;
+
+    @Override
+    public void redirectTo(URI uri)
+            throws RedirectException
+    {
+        redirectedTo = uri;
+    }
+
+    public URI redirectedTo()
+    {
+        return redirectedTo;
+    }
+}
